Fix can't change in-call volume with screen off
This is a fix for the following issue:
http://code.google.com/p/cyanogenmod/issues/detail?id=4070

Issue: Proximity sensor disables screen wake using power button. With the
"volume rocker wake" option enabled in CMParts, volume keys emulate the power
button whenever the screen is off. If the screen is off due to proximity
sensor, volume keys become useless.

Solution: When screen is off due to the proximity sensor, do NOT override the
volume keys as power buttons. This allows the user to adjust in-call volume
using volume keys while phone is to their ear.

Also: Small clean-up of unecessary code and style fixes.

Change-Id: I9f21bc1a59f940242b2196d423a7b8de62c855e0